addf?

+++ main.c
@@ -319,13 +319,13 @@ void handleMovRDLRs(CPU* cpu, uint8_t rd, uint8_t rs, uint64_t L) {
 void handleAddf(CPU* cpu, uint8_t rd, uint8_t rs, uint8_t rt) {
     double val1 = 0, val2 = 0;
     // Copy bits from the 64-bit register storage into double "number1" and "number2"
-    memcpy(&val1, &(cpu->registers[rs]), sizeof(double));
-    memcpy(&val2, &(cpu->registers[rt]), sizeof(double));
+    memcpy(&val1, &(cpu->registers[rs]), sizeof(uint64_t));
+    memcpy(&val2, &(cpu->registers[rt]), sizeof(uint64_t));
 
     double result = val1 + val2;
 
     // Copy the result back into register rd
-    memcpy(&(cpu->registers[rd]), &result, sizeof(double));
+    memcpy(&(cpu->registers[rd]), &result, sizeof(uint64_t));
 
     cpu->programCounter += 4;
 }
